+++ all-is-cubes/src/transaction.rs
@@ -34,7 +34,10 @@ pub use tester::*;
 /// If a transaction implements [`Default`], then the default value should be a
 /// transaction which has no effects and always succeeds, and is cheap to create.
 #[must_use]
-pub trait Transaction<T: ?Sized>: Merge {
+pub trait Transaction: Merge {
+    /// Type of the transaction’s target (what it can be used to mutate).
+    type Target;
+
     /// Type of a value passed from [`Transaction::check`] to [`Transaction::commit`].
     /// This may be used to pass precalculated values to speed up the commit phase,
     /// or even lock guards or similar, but also makes it slightly harder to accidentally
@@ -54,7 +57,7 @@ pub trait Transaction<T: ?Sized>: Merge {
     ///
     /// If the preconditions are met, returns [`Ok`] containing data to be passed to
     /// [`Transaction::commit`].
-    fn check(&self, target: &T) -> Result<Self::CommitCheck, PreconditionFailed>;
+    fn check(&self, target: &Self::Target) -> Result<Self::CommitCheck, PreconditionFailed>;
 
     /// Perform the mutations specified by this transaction. The `check` value should have
     /// been created by a prior call to [`Transaction::check()`].
@@ -69,12 +72,12 @@ pub trait Transaction<T: ?Sized>: Merge {
     ///
     /// The target should not be mutated between the call to [`Transaction::check()`] and
     /// [`Transaction::commit()`] (including via interior mutability, however that applies
-    /// to the particular `T`). The consequences of doing so may include mutating the
+    /// to the particular `Target` type). The consequences of doing so may include mutating the
     /// wrong components, signaling an error partway through the transaction, or merely
     /// committing the transaction while its preconditions do not hold.
     fn commit(
         &self,
-        target: &mut T,
+        target: &mut Self::Target,
         check: Self::CommitCheck,
         outputs: &mut dyn FnMut(Self::Output),
     ) -> Result<(), CommitError>;
@@ -96,7 +99,7 @@ pub trait Transaction<T: ?Sized>: Merge {
     /// See also: [`Transactional::transact()`], for building a transaction through mutations.
     fn execute(
         &self,
-        target: &mut T,
+        target: &mut Self::Target,
         outputs: &mut dyn FnMut(Self::Output),
     ) -> Result<(), ExecuteError<Self>> {
         let check = self.check(target).map_err(ExecuteError::Check)?;
@@ -108,10 +111,10 @@ pub trait Transaction<T: ?Sized>: Merge {
     /// so that it can be combined with other transactions in the same universe.
     ///
     /// This is a convenience wrapper around [`UTransactional::bind`].
-    fn bind(self, target: Handle<T>) -> UniverseTransaction
+    fn bind(self, target: Handle<Self::Target>) -> UniverseTransaction
     where
         Self: Sized,
-        T: UTransactional<Transaction = Self>,
+        Self::Target: UTransactional<Transaction = Self>,
     {
         UTransactional::bind(target, self)
     }
@@ -120,9 +123,8 @@ pub trait Transaction<T: ?Sized>: Merge {
 /// Merging two transactions (or, in principle, other values) to produce one result “with
 /// the effect of both”. Merging is a commutative, fallible operation.
 ///
-/// This is a separate trait from [`Transaction`] so that a single type can implement
-/// `Transaction<Foo>` and `Transaction<Bar>` without making it ambiguous which
-/// implementation `.merge()` refers to.
+/// This is a separate trait from [`Transaction`] because some components of transactions
+/// are mergeable but not executable in isolation.
 ///
 /// TODO: Generalize to different RHS types for convenient combination?
 pub trait Merge: Sized {
@@ -355,11 +357,11 @@ crate::util::cfg_should_impl_error! {
 
 /// Specifies a canonical [`Transaction`] type for the implementing type.
 ///
-/// [`Transaction<T>`](Transaction) may be implemented by multiple types but there can
-/// be at most one `<T as Transactional>::Transaction`.
+/// For a given `T`, [`Transaction<Target = T>`] may be implemented by multiple types,
+/// but there can be at most one `<T as Transactional>::Transaction`.
 pub trait Transactional {
     /// The type of transaction which should be used with `Self`.
-    type Transaction: Transaction<Self>;
+    type Transaction: Transaction<Target = Self>;
 
     /// Convenience method for building and then applying a transaction to `self`,
     /// equivalent to the following steps:
@@ -385,7 +387,7 @@ pub trait Transactional {
             &mut Self::Transaction,
             &Self,
         ) -> Result<O, <Self::Transaction as Merge>::Conflict>,
-        Self::Transaction: Transaction<Self, Output = NoOutput> + Default,
+        Self::Transaction: Transaction<Target = Self, Output = NoOutput> + Default,
     {
         let mut transaction = Self::Transaction::default();
         let output = f(&mut transaction, self).map_err(ExecuteError::Merge)?;