+++ CHANGELOG.md
@@ -71,6 +71,7 @@
     - `raytracer::SpaceRaytracer::trace_scene_to_text()` and `trace_scene_to_string()` have been replaced with the single method `to_text()`.
 
     - `transaction::ExecuteError` is now generic over the type of transaction it is an error from.
+    - `transaction::Transaction` now has an associated type `Target` instead of a type parameter; it is no longer possible for a transaction type to be used with more than one target type.
 
     - Renamed `universe::URef` to `universe::Handle`. Related items have also been renamed:
         - `URefErased` to `ErasedHandle`